refactor: separate bindings
